ep: Refactor NavigationSidepanel out of builder.ts
